Migrated to ament_cmake package. Avoided pytest-timeout dependency
