+++ launch_testing/dummy_tests/locking.cpp
@@ -1,4 +1,4 @@
- // Copyright 2016 Open Source Robotics Foundation, Inc.
+// Copyright 2019 Open Source Robotics Foundation, Inc.
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -12,7 +12,8 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-int main(int argc, char **argv) {
+int main()
+{
   while (1) {}
   return 0;
 }
