+++ launch_testing/test/launch_testing/test_gtest.py
@@ -15,23 +15,16 @@
 """Tests for the GTest Action."""
 
 import os
-import subprocess
 
 from launch import LaunchDescription
-from launch import LaunchIntrospector
 from launch import LaunchService
 from launch_testing.actions import GTest
 
 
 def test_gtest():
     """Test running a gtest with timeout."""
-    dir_path = os.path.dirname(os.path.realpath(__file__)) \
-        + '/../../dummy_tests'
-    path = dir_path + '/locking'
-    print(path)
-    subprocess.run(['cd ' + dir_path + '&& mkdir build && cd build '
-                    '&& cmake .. && make install && cd .. && rm -r build'],
-                   shell=True)
+    path = os.path.dirname(os.path.realpath(__file__)) \
+        + '/../../dummy_tests/locking'
     ld = LaunchDescription([
         GTest(
             path=path, timeout=5.0,
